refactor(story-maps): extract editor dashboard card
